Fix: Changed datatype from int to double to match that of the save file.

+++ FundMangler.cs
@@ -13,14 +13,14 @@ class FundMangler
 	{
 		string[] input = new string[1];
 		bool userFound = false;
-		int[] values;
+		double[] values;
 		List<string> budget = new List<string>();
 		if(File.Exists(filePath))
 		{
 			budget.AddRange(File.ReadAllLines(filePath));
 		}
 		int userAddress = -1;
-		values = new int[3]; 
+		values = new double[3]; 
 		
 		Console.WriteLine("Enter username:");
 		string user = Console.ReadLine();
@@ -44,11 +44,11 @@ class FundMangler
 			if(Console.ReadLine().ToLower() == "y")
 			{
 				Console.WriteLine("Input starting funds.");
-				values[0] = int.Parse(Console.ReadLine());
+				values[0] = double.Parse(Console.ReadLine());
 				Console.WriteLine("Input starting science.");
-				values[1] = int.Parse(Console.ReadLine());
+				values[1] = double.Parse(Console.ReadLine());
 				Console.WriteLine("Input starting reputation.");
-				values[2] = int.Parse(Console.ReadLine());
+				values[2] = double.Parse(Console.ReadLine());
 				
 				budget.Add(String.Format("{0},{1},{2},{3}", user, values[0], values[1], values[2]));
 				File.Create(filePath).Dispose();
@@ -71,7 +71,7 @@ class FundMangler
 				{
 					if(save[i].Trim().Split(' ')[0] == "funds")
 					{
-						values[0] = int.Parse(save[i].Trim().Split(' ')[2]);
+						values[0] = double.Parse(save[i].Trim().Split(' ')[2]);
 						break;
 					}
 				}
@@ -80,7 +80,7 @@ class FundMangler
 				{
 					if(save[i].Trim().Split(' ')[0] == "sci")
 					{
-						values[1] = int.Parse(save[i].Trim().Split(' ')[2]);
+						values[1] = double.Parse(save[i].Trim().Split(' ')[2]);
 						break;
 					}
 				}
@@ -89,7 +89,7 @@ class FundMangler
 				{
 					if(save[i].Trim().Split(' ')[0] == "rep")
 					{
-						values[2] = int.Parse(save[i].Trim().Split(' ')[2]);
+						values[2] = double.Parse(save[i].Trim().Split(' ')[2]);
 						break;
 					}
 				}
@@ -116,7 +116,7 @@ class FundMangler
 			case "l":
 				for(int i = 0; i < 3; i++)
 				{
-					values[i] = int.Parse(budget[userAddress].Split(',')[i + 1]);
+					values[i] = double.Parse(budget[userAddress].Split(',')[i + 1]);
 				}
 				
 				for(int i = FindBlockWithName("SCENARIO", "Funding"); i < save.Length; i++)
